Added Main Program File...
This will be the core file used for the project! I have all the basic parts set up to allow for the intake of string data. Will in a bit release another version that has a 2D double array which will contain all the data in one structure that follows the rules stated in the README.md file located in the root directory. Please take the time to read into the behavior of the program in order to fully understand what is going on. Make sure to use this file as the base program for when you're working on your individual parts so it may work correctly with the program.

Additional Notes:

I integrated Talha's data-entry code into the dataArrayCollection() function and made it more simplified to allow for less code to be used. I also created a function called arrayCharClean() which removes "junk" characters, such as quotation marks, and also denotes any blank entries with an '@' symbol for later use.

- Evan