Add files via upload
This is the modified version of the previous code I made. This code retrieves all the portions separated by a comma and converts the VALUE (Percentage) that we need for our calculation from a character pointer value to a float value. We do not need to use this code, I was just playing around to familiarize how to use strings.