Added Information on New Arrays...
Updated the README.md to included an update made to the "Important Details" section of the program. This file now includes all information on the new sorting and saving system created with the use of the new array variables, along with behaviour information. This will help in understanding how the program works now since being heavily re-optimized by myself!

- Evan Shizas