Major Changes Made to Program!!
This upload is a major modification of what was previously listed on this branch! Please take the time to read and view the changes made in this push! This will be key for when working on the report, as the behaviour has changed drastically in some areas enough that the reasoning behind how certain functions work may not be as it seems!

The "CPS188_Term_Project.c" file is now fully finished! I have implemented everyone's work, along with my own, into the main program. The results coming back from the program are one-to-one with those of the work done in the user-specific files, so all is good with regards to that. All major changes done to the program were all in the name of optimization, along with removal of hard/repeated-coding. With the changes made to the structure of the program, I have cut the size of the file by more than 650 lines! 

So, the major changes done to the program apply to how the program both saves and calls back variables containing key data! I noticed that when the code was first being implemented, a lot of variables existed that had practically the same purpose as one-another, with the only difference being what specific data they stored for certain locations/age-groups. With that knowledge in mind, I went ahead and created three new arrays called avgData[], avgYearData[][], and avgAgeData[][]. These variables store all the data/calculations asked for in the term project's instructions, just like the previous variables it replaced, but in a more convenient, less crude way. This new storage behaviour - which is spoken about more in the README.md - allowed for simple for-loops to do a lot of the heavy-lifting of collecting, calculating and displaying of data. This in-turn allowed for less repetitive code to exist that Reggie may believe to be hard-coding; costing us marks! 

On another note, the program is now also more flexible in terms of varying conditions. This allows for the system to adapt to changes made by the end-user which could crash the program when executing. Such examples include having a lot of static variables predefined using the "#define" macro, as well as file scanner functions that see how large a data set is in order to determine the sizing of array variables. These can then be changed by the user to accommodate their needs. In addition to these changes, I have also made sure to include Jonathan's file-writing system and am certain that the written text file created by the main program was exactly the same to that of his work (same name and printing structure); this should prevent any issues with the GNUPlot scripts written by Fergus. 

In the end of it all, the work was really well done by everyone! Major props on getting it done relatively quick as well!

Feel free to try out the program and let me know if you would like any further explanations of what was done, along with any last minute changes you may have in mind.

- Evan Shizas